feat: cross-platform ARM bare-metal support for CMSIS-OS backend
- osal_thread_stop.h: replace thread_local with FreeRTOS TLS slot 0 on
  DPSDK_BARE_METAL (pvTaskGetThreadLocalStoragePointer / vTaskSetThreadLocal-
  StoragePointer). POSIX/desktop path unchanged (still uses thread_local).
  Requires configNUM_THREAD_LOCAL_STORAGE_POINTERS >= 1 in FreeRTOSConfig.h.
- osal_thread.h: update two call sites from tl_cmsis_stop_flag = ... to
  osal_cmsis_stop_flag_set(...) using the new cross-platform API.
- osal_system.h: update sleep_ms stop-check to use osal_cmsis_stop_flag_get().
- osal_thread_pool.cpp: static_cast<unsigned int> for %u printf args
  (uint32_t = unsigned long on ARM; fixes -Werror=format on arm-none-eabi-gcc).

+++ src/impl/cmsis_os/osal_thread_pool.cpp
@@ -28,7 +28,7 @@ void OSALThreadPool::start(uint32_t numThreads, int priority, int stack_size) {
     for (uint32_t i = 0; i < numThreads; ++i) {
         OSALThreadPool::OSALAddTread();
     }
-    OSAL_LOGD("Thread pool started with %u threads\n", numThreads);
+    OSAL_LOGD("Thread pool started with %u threads\n", static_cast<unsigned int>(numThreads));
 }
 
 bool OSALThreadPool::OSALAddTread() {
@@ -146,14 +146,14 @@ void OSALThreadPool::setTaskFailureCallback(std::function<void(void *)> callback
 
 void OSALThreadPool::setMaxThreads(uint32_t maxThreads) {
     maxThreads_ = maxThreads;
-    OSAL_LOGD("Max threads set to %u\n", maxThreads);
+    OSAL_LOGD("Max threads set to %u\n", static_cast<unsigned int>(maxThreads));
 }
 
 uint32_t OSALThreadPool::getMaxThreads() const { return maxThreads_; }
 
 void OSALThreadPool::setMinThreads(uint32_t minThreads) {
     minThreads_ = minThreads;
-    OSAL_LOGD("Min threads set to %u\n", minThreads);
+    OSAL_LOGD("Min threads set to %u\n", static_cast<unsigned int>(minThreads));
 }
 
 uint32_t OSALThreadPool::getMinThreads() const { return minThreads_; }
